Fbx v2

+++ src/MeshIO/Utils.cs
@@ -16,13 +16,13 @@ namespace MeshIO
 		/// Creates an id as a long.
 		/// </summary>
 		/// <returns></returns>
-		public static ulong CreateId()
+		public static long CreateId()
 		{
 			lock (_syncLock)
 			{
 				byte[] buffer = new byte[8];
 				_random.NextBytes(buffer);
-				return (ulong)Math.Abs(BitConverter.ToInt64(buffer, 0));
+				return Math.Abs(BitConverter.ToInt64(buffer, 0));
 			}
 		}
 	}
